requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
PyJWT==2.8.0
pytz==2023.4
requests==2.31.0
sqlparse==0.4.4
twilio==8.12.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4